Skyline v0.0.1 Skyline.Amnesia.Topic.Database

Functions

create(copying \\ [])

Specs

create(Amnesia.Table.c) :: [Amnesia.Table.o]

Create the database, it calls .create on every defined table.

create!(copying \\ [])

Specs

create!(Amnesia.Table.c) :: [Amnesia.Table.o]

Create the database, it calls .create on every defined table, raises in case of error.

destroy()

Specs

destroy :: [Amnesia.Table.o]

Destroy the database, it calls .destroy on every defined table.

destroy!()

Specs

destroy! :: [Amnesia.Table.o]

Destroy the database, it calls .destroy on every defined table, raises in case of error.

metadata()

Specs

metadata :: Amnesia.Metadata.t

tables()

Specs

tables :: [atom]

List of atoms of the defined tables in the database.

wait(timeout \\ :infinity)

Specs

wait(integer | :infinity) ::
  :ok |
  {:timeout, [atom]} |
  {:error, atom}

Wait for the database to be loaded.

Macros

__using__(opts)

Alias all the table names in the current scope and require what’s needed.