Skyline v0.0.1 Skyline.Topic.Dispatcher

Router

A routing scheme inspired by emqttd.

This tree differs from emqttd by acting as a registry for the pids themselves instead of acting as a registry for active topics, which are then dispatched seperately.

It also differs by each key being immutable and tagged by pid of the subscriber process. This makes write locks unecassary and lets you remove subscriptions and prune leaves without walking the tree. In emqttd trees, the nodes count the leaves by the value on the edge which gets updated.

No benchmarks have been done to see if one is more efficient then the other.

Functions

add_topic_subscription(topic, pid)

Specs

add_topic_subscription(String.t, pid) :: :ok

Register a subcsriber pid with a topic

broadcast_msg(topic, msg)

Specs

broadcast_msg(String.t, any) :: :ok

Cast a message to all pid’s registered with the topic

prune_tree(pid)

Specs

prune_tree(pid) :: :ok

Remove all edges associated with pid