Skyline v0.0.1 Skyline.Session

Responsible for writing messages to a socket, starting up new subcriptions, and managing the session.

Functions

send_to_socket(socket, msg)

Specs

send_to_socket(Skyline.socket, Skyline.skyline_msg) :: :ok

An exposed helper method for sending a message to a socket

start_link(socket, con, opts \\ [])

Specs

start_link(Skyline.socket, Skyline.Msg.Connect.t, [{:key, any}]) :: GenServer.on_start