skyline v0.0.1 Skyline.Topic.Controller

Implements the controller pipeline responsible for handling publish and subscribe messages.

The pipeline

The goal of a controller is to receive a request and invoke the desired action. The whole flow of the controller is managed by a single pipeline:

defmodule UserController do
  use Skyline.Topic.Controller
  require Logger
  pipe :log_message, "before action"
  def publish(conn, _params) do
    Logger.debug "publish/2"
    conn
  end
  defp log_message(conn, msg) do
    Logger.debug msg
    conn
  end
end

When invoked, this pipeline will print:

before action
show/2

As any other pipeline, we can halt at any step by returning {:close_connection, reason} where reason can be an exception or a string. If we change log_message/2 to:

def log_message(conn, msg) do
  Logger.debug msg
  {:close_connection, "Halt pipe"}
end

it will print only:

before action

As the rest of the pipeline (the action and the after action pipe) will never be invoked.

Guards

pipe/2 supports guards, allowing a developer to configure a pipe to only run in some particular action:

pipe :log_message, "before publish" when action == :publish

The first pipe will run only when action is publish. Those guards work like regular Elixir guards and the only variables accessible in the guard are conn, the action as an atom and the controller as an alias.

Controllers are pipes

Like routers, controllers are pipes, but they are wired to dispatch to a particular function which is called an action. For example, the route:

subscribe "/users/:id", UserController

will invoke UserController as a pipe:

UserController.call(conn, :subscribe)

which will trigger the pipe pipeline and which will eventually invoke the inner action pipe that dispatches to the show/2 function in the UserController. As controllers are pipes, they implement both init/1 and call/2, and it also provides a function named action/2 which is responsible for dispatching the appropriate action after the pipe stack (and is also overridable).

Macros

pipe(pipe)

Stores a pipe to be executed as part of the pipe pipeline.

pipe(pipe, opts)

Stores a pipe with the given options to be executed as part of the pipe pipeline.