skyline v0.0.1 Skyline.Topic.Pipe behaviour

The pipe specification for transforming Skyline.Conn.

There are two kind of pipes: function pipes and module pipes.

Function pipes

A function pipe is any function that receives a connection and a set of options and returns a connection. Its type signature must be:

(Skyline.Conn.t, Plug.opts) :: Plug.Conn.t

Module pipes

A module pipe is an extension of the function pipe. It is a module that must export:

a call/2 function wipipeth the signature defined above

an init/1 function which takes a set of options and initializes it. The result returned by init/1 is passed as second argument to call/2. Note that init/1 may be called during compilation and as such it must not return pids, ports or values that are not specific to the runtime. The API expected by a module pipe is defined as a behaviour by the pipe module (this module).

Examples

Here’s an example of a function pipe:

def lowercase_topic(%Conn{topic: topic} = conn, opts) do
  %{conn | topic: String.downcase(topic)}
end

Here’s an example of a module pipe:

defmodule LowercaseTopicPipe do
  def init(opts) do
    opts
  end
  def call(%Conn{topic: topic} = conn, _opts) do
     %{conn | topic: String.downcase(topic)}
  end
end

Types

opts :: tuple | atom | integer | float | [opts]

pipe :: module | atom

response ::
  :ok |
  {:close_connection, String.t} |
  :do_nothing |
  {:topic_qos, Skyline.qos_type}

Functions

compile(env, pipeline, builder_opts)

Specs

compile(Macro.Env.t, [{pipe, Skyline.Topic.Pipe.opts, Macro.t}], Keyword.t) :: {Macro.t, Macro.t}

Callbacks

call(arg0, term)

Specs

call(Skyline.Conn.t, term) :: response

init(term)

Specs

init(term) :: term