skyline v0.0.1 Skyline.Topic.Router

Macros

match(verb, path, pipe, pipe_opts, options \\ [])

pipe(pipe, opts \\ [])

pipe_through(pipes)

pipeline(pipe, list)

publish(path, pipe, options \\ [])

Generates a route to handle a publish request to the given path.

scope(options, list)

scope(path, options, list)

Define a scope with the given path. This function is a shortcut for:

scope path: path do
  ...
end

scope(path, alias, options, list)

Defines a scope with the given path and alias. This function is a shortcut for:

scope path: path, alias: alias do
  ...
end

subscribe(path, pipe, options \\ [])

Generates a route to handle a subscribe request to the given path.