skyline v0.0.1 Skyline.Topic.Utils

Utility functions for building Publish and Subscribe Controllers.

Functions

publish(topic, qos, msg, client)

Specs

publish(String.t, Skyline.qos_type, Skyline.Msg.PublishReq.t, Skyline.Client.t) :: Skyline.Client.t

Deliver the message to all subscribers, retain if necassary.

subscribe(topic, qos, subscribe, client)

Specs

subscribe(String.t, Skyline.qos_type, Skyline.Msg.Subscribe.t, Skyline.Client.t) :: Skyline.Client.t

Sets up a subsription to the topic and register with tree dispatcher.