skyline v0.0.1 Skyline.Topic.Router

Defines a Skyline router.

Skyline router is a simplified version of Phoenix’s so don’t be surprised if there are simularities. The router provides a set of macros for generating routes that dispatch to specific controllers and actions. Those macros are named after message types publish and subscribe. For example:

defmodule MyApp.Router do
  use Skyline.Topic.Router
  publish "/topic/:topic", PublishController
end

The publish/2 macro above accepts a request of format "/topic/VALUE" and dispatches it to PublishController.publish/2. Routes can also match glob-like patterns, routing any path with a common base to the same controller. For example:

publish "/dynamic*anything", DynamicController

Skyline’s router is extremely efficient, as it relies on Elixir pattern matching for matching routes and serving requests.

Scopes and Resources

The router also supports scoping of routes:

scope "/api/v1" do
  subscribe "/devices/:id", DeviceController
end

For example, the route above will match on the path "/api/v1/devices/:id" and the named route will be api_v1_page_path, as expected from the values given to scope/2 option. Skyline also provides a resource/2 macro that allows developers to shortcut publish/3 and subscribe/3:

defmodule MyApp.Router do
  use Skyline.Topic.Router
  resource "/devices", DeviceController
  resource "/users", UserController
end

Check scope/2 and resource/2 for more information.

Pipelines and pipes

Once a request arrives at the Skyline router, it performs a series of transformations through pipelines until the request is dispatched to a desired end-point. Such transformations are defined via pipes, as defined in Skyline.Pipe. Once a pipeline is defined, it can be piped through per scope. For example:

defmodule MyApp.Router do
  use Skyline.Topic.Router
  pipeline :session do
    pipe :fetch_session
  end
  scope "/" do
    pipe_through :session
    # session related routes and resources
  end
end

Skyline.Topic.Router imports functions from both Skyline.Conn and Skyline.Controller to help define pipes. In the example above, fetch_session/2 comes from Skyline.Conn while accepts/2 comes from Skyline.Controller. Note that router pipelines are only invoked after a route is found. No pipe is invoked in case no matches were found.

Macros

match(verb, path, controller, pipe_opts, options \\ [])

pipe(pipe, opts \\ [])

pipe_through(pipes)

pipeline(pipe, list)

publish(path, controller, options \\ [])

Generates a route to handle a publish request to the given path.

resource(path, pipe, options \\ [])

scope(options, list)

scope(path, options, list)

Define a scope with the given path. This function is a shortcut for:

scope path: path do
  ...
end

scope(path, alias, options, list)

Defines a scope with the given path and alias. This function is a shortcut for:

scope path: path, alias: alias do
  ...
end

subscribe(path, controller, options \\ [])

Generates a route to handle a subscribe request to the given path.